Add DSPBase and DSP7225
