+++ docs/api/instruments/signalrecovery/dsp7265.rst
@@ -5,3 +5,4 @@ DSP 7265 Lock-in Amplifier
 .. autoclass:: pymeasure.instruments.signalrecovery.DSP7265
     :members:
     :show-inheritance:
+    :inherited-members: CommonBase, Instrument
